+++ 20_ETH_serwer_www_test_step+lcd/step.c
@@ -13,8 +13,8 @@
 #include "step.h"
 
 uint8_t start_stepper = 0;
-int8_t steps_state = 0;
-uint8_t steps_received = 0;
+int16_t steps_state = 0;
+int8_t steps_received = 0;
 int8_t steps_todo = 0;
 int8_t oversteps = 0;
 uint8_t steps_cmd = 0;
